Modularized all of the code (and it works I think)
