+++ lib/designateMixingBridge.js
@@ -1,6 +1,7 @@
 var Q = require('q');
 var isStation = require('./isStation.js');
 var updateState = require('./updateState.js');
+var customError = require('./customError.js');
 
 /**
  * Looks for an existing mixing bridge and creates one if none exist.
@@ -72,19 +73,19 @@ var designateMixingBridge = function(client, extsInUse, bridgeName, states) {
    * @param {Object} object - contains the bridge and channel in question
    */
   function channelLeftBridge(event, object) {
-    defer = Q.defer();
     var extension = extsInUse[bridgeName];
     if (isStation(extension, object.channel)) {
+      extsInUse[bridgeName].removeStation(object.channel.name);
       if (object.bridge.channels.length === 1) {
         var hangupTrunk = Q.denodeify(client.channels.hangup.bind
             (client.channels));
         hangupTrunk({channelId: object.bridge.channels[0]})
           .catch(function (err) {
             err.name = 'HangupFailure';
-            defer.reject(err);
           });
       }
     } else {
+      extsInUse[bridgeName].removeTrunk();
       if (object.bridge.channels.length) {
         console.log('Hanging up station channel/s');
         var hangupArray = [];
@@ -98,8 +99,8 @@ var designateMixingBridge = function(client, extsInUse, bridgeName, states) {
     }
 
     if (object.bridge.channels.length === 0) {
-      defer.resolve();
       updateState(client, extsInUse, bridge.name, states.IDLE);
+      extsInUse.freeExtension(bridgeName);
       bridge.removeAllListeners('ChannelEnteredBridge');
       bridge.removeAllListeners('ChannelLeftBridge');
     }
